Enable completion through CompleteDone event
Support an alternative solution for completion without the need to use
an extra completion item selection mapping. Therefore the user can
establish an autocommand called listen to the CompleteDone event. There
it will be checked if the completed item was from kind function and then
the parameter gets completed. Therefore the returned key sequence must
be feed manually by the autocommand, in contrast to the mapping where
the keys would get executed naturally.
This commit introduce only a new variable to enable this alternative by
disabling the check if the pop-up menu is visible on calling the
completion. Since this is not the case for the previous presented
solution, it must be possible to disable this.

Related to: tenfyzhong/CompleteParameter.vim#44